pyproject.toml
[project]
name = "text-to-sql-agent"
version = "0.1.0"
description = "A Text-to-SQL agent using MySQL, RAG, and DeepSeek"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "openai>=1.0.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "chromadb>=0.4.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pandas>=2.0.0",
    "tabulate>=0.9.0",
    "cryptography>=42.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "sqlglot>=18.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
